[README] Update the new features section. [Guide] Add navigation and optimization modpacks section [Language] (mcl_memory_allocation) Change string from "Memory allocation" to "Maximum memory allocation" [Language] (mcl_memory_allocation_subtitle) Change string from "Controls how much memory is given to Minecraft" to "Controls the maximum amount of memory Minecraft can use." [Settings] (Java Tweaks) Add Minimum memory allocation slider [Settings] (Java Tweaks) Set minimum amount of RAM to 128 MB [Settings] (Java Tweaks) Add option to disable purging of certain JVM arguments.

+++ app_pojavlauncher/src/main/java/net/kdt/pojavlaunch/utils/JREUtils.java
@@ -293,15 +293,22 @@ public class JREUtils {
         List<String> userArgs = getJavaArgs(activity);
 
         //Remove arguments that can interfere with the good working of the launcher
-        purgeArg(userArgs,"-Xms");
-        purgeArg(userArgs,"-Xmx");
-        purgeArg(userArgs,"-d32");
-        purgeArg(userArgs,"-d64");
-        purgeArg(userArgs, "-Xint");
-        purgeArg(userArgs, "-Dorg.lwjgl.opengl.libname");
+        if (LauncherPreferences.PREF_ENABLE_JVM_ARGUMENTS_PURGER) {
+            Logger.getInstance().appendToLog("JVM arguments purger enabled!");
+            purgeArg(userArgs,"-Xms");
+            purgeArg(userArgs,"-Xmx");
+            purgeArg(userArgs,"-d32");
+            purgeArg(userArgs,"-d64");
+            purgeArg(userArgs, "-Xint");
+            purgeArg(userArgs, "-Dorg.lwjgl.opengl.libname");
+        }
+        else {
+            // Don't do anything
+            Logger.getInstance().appendToLog("JVM arguments purger disabled (This may cause problems).");
+        }
 
         //Add automatically generated args
-        userArgs.add("-Xms" + LauncherPreferences.PREF_RAM_ALLOCATION + "M");
+        userArgs.add("-Xms" + LauncherPreferences.PREF_MINIMUM_RAM_ALLOCATION + "M");
         userArgs.add("-Xmx" + LauncherPreferences.PREF_RAM_ALLOCATION + "M");
         if(LOCAL_RENDERER != null) userArgs.add("-Dorg.lwjgl.opengl.libname=" + graphicsLib);
